# Тили-тили/Тили-тили_API_openapi.yaml
openapi: 3.0.3
info:
  title: Тили-тили API
  version: 0.28.0
  description: |
    Контракт бэкенда для PWA «Тили-тили» (tili-tili.ru) — свадебный планировщик.
    Фронтенд (React) уже работает на моках; этот файл — единственный источник правды для бэкенда (Claude Code).

    ## Решения владельца 2026-09-02 (контракт v0.2)
    - Деньги — объект `{ amount: копейки, currency: ISO 4217 }` у КАЖДОЙ суммы.
      Не одна валюта на свадьбу: план §19.7 описывает свадьбу за границей, где
      сделка в евро, а бюджет в рублях. В MVP принимается только RUB.
    - У слота нет собственного статуса. Слот либо пуст, либо несёт сделку;
      состояние живёт в `Deal.state` — шесть значений. `Slot.tileState` —
      производная подпись для мозаики, только для чтения.
    - Гость опознаётся ОДНИМ персональным токеном во всех гостевых путях
      (`/rsvp/{guestToken}`, `/gifts/{guestToken}`, `/join/{guestToken}/…`).
      Общий код свадьбы не годится: по нему нельзя подставить имя в приглашение
      и нельзя дать гостю снять СВОЙ резерв подарка. Анонимность сохраняется —
      система знает гостя, а API пары этот токен не отдаёт никогда (§9).
    - Переноса данных с localStorage-моков не будет: реальных пользователей не
      было, эндпоинт импорта не заводится.

    ## Роли
    - `couple` — молодожёны (владельцы свадьбы): полный доступ.
    - `helper` — помощник (друг/родственник): орг-разделы, без бюджета и сделок.
...
    - `vendor` — подрядчик: свои сделки, лиды, чаты, анкета.
...

...
servers:
...
paths:
...
  /weddings/{weddingId}/timeline/shift:
    post:
      tags: [timeline]
      summary: Сдвинуть день X на N минут
      description: 'Кнопка «+15 мин» в режиме дня X. Сдвигает все последующие блоки и рассылает push команде и гостям — killer-фича §19.6.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WeddingId' }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [minutes]
              properties:
                minutes: { type: integer, example: 15 }
      responses:
        '200': { description: OK }
  /weddings/{weddingId}/planb:
    get:
      tags: [timeline]
      summary: "План Б: сценарий и чек-лист накануне"
      description: >
        §13.1 называет чек-лист персистентным: отметки переживают смену
        устройства и видны обоим партнёрам. Пункты — обычные задачи вида
        `planb`, поэтому отмечаются через PATCH /weddings/{weddingId}/tasks/{taskId}
        и в общий чек-лист по месяцам не попадают.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/WeddingId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenario:
                    type: string
                    nullable: true
                    description: 'Объявленный сценарий или null, если план Б не активировали.'
                  activatedAt: { type: string, format: date-time, nullable: true }
                  checklist:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        title: { type: string }
                        done: { type: boolean }
  /weddings/{weddingId}/planb/activate:
    post:
      tags: [timeline]
      summary: Активировать план Б
      description: 'Фиксирует запасной сценарий и время его активации и уведомляет команду в приложении — проверить точку сбора и тайминг (§13.1). Тайминг не пересобирается, гостям ничего не рассылается: канала доставки гостям пока нет.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/WeddingId' }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenario]
              properties:
                scenario: { type: string, example: rain }
      responses:
        '200': { description: OK }

  # ───────────────────────── ВДОХНОВЕНИЕ (§8.4) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WeddingId:
      name: weddingId
      in: path
      required: true
      schema: { type: string }
...
security:
  - bearerAuth: []